+++ CHANGELOG.md
@@ -5,3 +5,4 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [0.1.0](https://github.com/omnibenchmark/omni-py) (unreleased)
 - Setup CI
+- Add first version of remote storage backend